Refactored package structure for sanity sources
