+++ src/main/java/croissantnova/sanitydim/sources/passive/BlockStuck.java
@@ -1,4 +1,4 @@
-package croissantnova.sanitydim.passive;
+package croissantnova.sanitydim.sources.passive;
 
 import croissantnova.sanitydim.capability.IPersistentSanity;
 import croissantnova.sanitydim.capability.ISanity;
