+++ src/main/java/croissantnova/sanitydim/sources/passive/Darkness.java
@@ -1,4 +1,4 @@
-package croissantnova.sanitydim.passive;
+package croissantnova.sanitydim.sources.passive;
 
 import croissantnova.sanitydim.capability.ISanity;
 import croissantnova.sanitydim.config.ConfigProxy;
@@ -7,11 +7,14 @@ import net.minecraft.server.level.ServerPlayer;
 
 import javax.annotation.Nonnull;
 
-public class Passive implements IPassiveSanitySource
+public class Darkness implements IPassiveSanitySource
 {
     @Override
     public float get(@Nonnull ServerPlayer player, @Nonnull ISanity cap, @Nonnull ResourceLocation dim)
     {
-        return ConfigProxy.getPassive(dim);
+        if (player.level().getMaxLocalRawBrightness(player.blockPosition()) <= ConfigProxy.getDarknessThreshold(dim))
+            return ConfigProxy.getDarkness(dim);
+
+        return 0;
     }
 }
