+++ src/main/java/croissantnova/sanitydim/sources/passive/Passive.java
@@ -1,4 +1,4 @@
-package croissantnova.sanitydim.passive;
+package croissantnova.sanitydim.sources.passive;
 
 import croissantnova.sanitydim.capability.ISanity;
 import croissantnova.sanitydim.config.ConfigProxy;
@@ -7,14 +7,11 @@ import net.minecraft.server.level.ServerPlayer;
 
 import javax.annotation.Nonnull;
 
-public class Darkness implements IPassiveSanitySource
+public class Passive implements IPassiveSanitySource
 {
     @Override
     public float get(@Nonnull ServerPlayer player, @Nonnull ISanity cap, @Nonnull ResourceLocation dim)
     {
-        if (player.level().getMaxLocalRawBrightness(player.blockPosition()) <= ConfigProxy.getDarknessThreshold(dim))
-            return ConfigProxy.getDarkness(dim);
-
-        return 0;
+        return ConfigProxy.getPassive(dim);
     }
 }
